+++ features/fixtures/maze_runner/Assets/Scripts/MobileScenarioRunner.cs
@@ -38,6 +38,8 @@ public class MobileScenarioRunner : MonoBehaviour {
         {"21", "Java Background Crash No Threads" },
         {"22", "iOS Native Error" },
         {"23", "iOS Native Error No Threads" },
+        {"24", "Mark Launch Complete" },
+        {"25", "Check Last Run Info" },
 
 
 
@@ -186,6 +188,13 @@ public class MobileScenarioRunner : MonoBehaviour {
     {
         switch (scenarioName)
         {
+            case "Check Last Run Info":
+                CheckLastRunInfo();
+                break;
+            case "Mark Launch Complete":
+                Bugsnag.MarkLaunchCompleted();
+                ThrowException();
+                break;
             case "Android Persistence Directory":
                 CheckForCustomAndroidDir();
                 break;
@@ -264,6 +273,15 @@ public class MobileScenarioRunner : MonoBehaviour {
                 break;
             default:
                 throw new System.Exception("Unknown scenario: " + scenarioName);
+        }
+    }
+
+    private void CheckLastRunInfo()
+    {
+        var info = Bugsnag.GetLastRunInfo();
+        if (info.Crashed && info.CrashedDuringLaunch && info.ConsecutiveLaunchCrashes > 0)
+        {
+            Bugsnag.Notify(new System.Exception("Last Run Info Correct"));
         }
     }