+++ src/BugsnagUnity/Native/Android/NativeInterface.cs
@@ -13,6 +13,7 @@ namespace BugsnagUnity
         private IntPtr BugsnagNativeInterface;
         private IntPtr BugsnagUnityClass;
         // Cache of classes used:
+        private IntPtr LastRunInfoClass;
         private IntPtr BreadcrumbClass;
         private IntPtr BreadcrumbTypeClass;
         private IntPtr CollectionClass;
@@ -83,6 +84,10 @@ namespace BugsnagUnity
                 IntPtr crumbRef = AndroidJNI.FindClass("com/bugsnag/android/Breadcrumb");
                 BreadcrumbClass = AndroidJNI.NewGlobalRef(crumbRef);
                 AndroidJNI.DeleteLocalRef(crumbRef);
+
+                IntPtr lastRunInfoRef = AndroidJNI.FindClass("com/bugsnag/android/LastRunInfo");
+                LastRunInfoClass = AndroidJNI.NewGlobalRef(lastRunInfoRef);
+                AndroidJNI.DeleteLocalRef(lastRunInfoRef);
 
                 IntPtr crumbTypeRef = AndroidJNI.FindClass("com/bugsnag/android/BreadcrumbType");
                 BreadcrumbTypeClass = AndroidJNI.NewGlobalRef(crumbTypeRef);
@@ -190,6 +195,8 @@ namespace BugsnagUnity
             obj.Call("setMaxBreadcrumbs", config.MaximumBreadcrumbs);
             obj.Call("setMaxPersistedEvents", config.MaxPersistedEvents);
             obj.Call("setPersistUser", config.PersistUser);
+            obj.Call("setLaunchDurationMillis", config.LaunchDurationMillis);
+            obj.Call("setSendLaunchCrashesSynchronously", config.SendLaunchCrashesSynchronously);
 
             // set endpoints
             var notify = config.Endpoints.Notify.ToString();
@@ -376,6 +383,11 @@ namespace BugsnagUnity
           session.HandledCount()
         });
             }
+        }
+
+        public void MarkLaunchCompleted()
+        {
+            CallNativeVoidMethod("markLaunchCompleted", "()V", new object[] {});
         }
 
         public Dictionary<string, object> GetApp()
@@ -815,5 +827,21 @@ namespace BugsnagUnity
 
             return dict;
         }
+
+        public LastRunInfo GetlastRunInfo()
+        {
+            var javaLastRunInfo = CallNativeObjectMethodRef("getLastRunInfo", "()Lcom/bugsnag/android/LastRunInfo;", new object[] { });
+            var crashed = AndroidJNI.GetBooleanField(javaLastRunInfo , AndroidJNIHelper.GetFieldID(LastRunInfoClass,"crashed"));
+            var consecutiveLaunchCrashes = AndroidJNI.GetIntField(javaLastRunInfo, AndroidJNIHelper.GetFieldID(LastRunInfoClass, "consecutiveLaunchCrashes"));
+            var crashedDuringLaunch = AndroidJNI.GetBooleanField(javaLastRunInfo, AndroidJNIHelper.GetFieldID(LastRunInfoClass, "crashedDuringLaunch"));
+            var lastRunInfo = new LastRunInfo
+            {
+                ConsecutiveLaunchCrashes = consecutiveLaunchCrashes,
+                Crashed = crashed,
+                CrashedDuringLaunch = crashedDuringLaunch
+            };
+            return lastRunInfo;
+        }
+
     }
 }